% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginv.R
\name{ginv}
\alias{ginv}
\title{finding generalized inverse of a matrix}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{matrix}

\item{tol}{tolerance}
}
\value{
a matrix
}
\description{
finding generalized inverse of a matrix
}
