% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearDat.R
\docType{data}
\name{nearDat}
\alias{nearDat}
\title{A simulated data that has near to separation problem}
\format{
\subsection{\code{nearDat}}{

A data frame with 120 rows and 4 columns:
\describe{
\item{id}{clusted id}
\item{y}{disease status}
\item{x}{treatment status}
\item{obstime}{follow-up time}
}
}
}
\usage{
nearDat
}
\description{
A simulated data that has near to separation problem
}
\keyword{datasets}
