% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiDat.R
\docType{data}
\name{quasiDat}
\alias{quasiDat}
\title{A simulated data that has near to separation problem}
\format{
\subsection{\code{quasiDat}}{

A data frame with 120 rows and 4 columns:
\describe{
\item{id}{clusted id}
\item{y}{disease status}
\item{x}{treatment status}
\item{obstime}{follow-up time}
}
}
}
\usage{
quasiDat
}
\description{
A simulated data that has near to separation problem
}
\keyword{datasets}
