% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xch.R
\name{xch}
\alias{xch}
\title{finding intra-cluster correlation matrix}
\usage{
xch(n, rho)
}
\arguments{
\item{n}{cluster size}

\item{rho}{icc}
}
\value{
a matrix
}
\description{
xch for exchangeable and ar1 for autoregressive 1
}
